Move fit_energy_functions from core to NB_utils
