+++ krcal/NB_utils/fit_energy_functions.py
@@ -1,0 +1,184 @@
+"""Module fit_energy_functions.
+This module includes the fit functions for the energy.
+
+Notes
+-----
+    KrCalib code depends on the IC library.
+    Public functions are documented using numpy style convention
+
+Documentation
+-------------
+    Insert documentation https
+"""
+import numpy as np
+from typing      import List, Tuple, Iterable
+import warnings
+
+from   invisible_cities.core.core_functions import in_range
+import invisible_cities.core.fit_functions as fitf
+
+from   invisible_cities.evm  .ic_containers  import Measurement
+from ..core. fit_functions import chi2
+from .. core. stat_functions import mean_and_std
+
+from ..core. kr_types import Number, Range
+
+from invisible_cities.core .stat_functions import poisson_sigma
+from invisible_cities.icaro. hst_functions import shift_to_bin_centers
+from invisible_cities.types.ic_types       import NN
+
+from ..core. kr_types import GaussPar
+from ..core. kr_types import FitPar
+from ..core. kr_types import FitResult
+from ..core. kr_types import HistoPar
+from ..core. kr_types import FitCollection
+
+
+from scipy.optimize import OptimizeWarning
+from numpy import sqrt, pi
+
+
+
+def gfit(x     : np.array,
+         y     : np.array,
+         yu    : np.array,
+         fseed : Tuple[float, float, float]) ->Tuple[FitPar, FitResult]:
+
+    f     = fitf.fit(fitf.gauss, x, y, fseed, sigma=yu)
+    c2    = chi2(f, x, y, yu)
+    par  = np.array(f.values)
+    err  = np.array(f.errors)
+    xu   = np.diff(x) * 0.5
+
+    fr = FitResult(par = par,
+                   err = err,
+                   chi2 = c2,
+                   valid = True)
+    fp = FitPar(x  = x, y  = y, xu = xu, yu = yu, f  = f.fn)
+
+    return fp, fr
+
+def gaussian_parameters(x : np.array, range : Tuple[Number], bin_size : float = 1)->GaussPar:
+    """
+    Return the parameters defining a Gaussian
+    g = N * exp(x - mu)**2 / (2 * std**2)
+    where N is the normalization: N = 1 / (sqrt(2 * pi) * std)
+    The parameters returned are the mean (mu), standard deviation (std)
+    and the amplitude (inverse of N).
+    """
+    mu, std = mean_and_std(x, range)
+    ff     = sqrt(2 * pi) * std
+
+    amp     = len(x) * bin_size / ff
+
+    sel  = in_range(x, *range)
+    N = len(x[sel])              # number of samples in range
+    mu_u  = std / sqrt(N)
+    std_u = std / sqrt(2 * (N -1))
+    amp_u = sqrt(2 * np.pi) * std_u
+
+    return GaussPar(mu  = Measurement(mu, mu_u),
+                    std = Measurement(std, std_u),
+                    amp = Measurement(amp, amp_u))
+
+
+def gaussian_fit(x       : np.array,
+                 y       : np.array,
+                 seed    : GaussPar,
+                 n_sigma : int)  ->Tuple[FitPar, FitResult]:
+    """Gaussian fit to x,y variables, with fit range defined by n_sigma"""
+
+    mu  = seed.mu.value
+    std = seed.std.value
+    amp = seed.amp.value
+    fit_range = mu - n_sigma * std, mu + n_sigma * std
+
+    x, y      = x[in_range(x, *fit_range)], y[in_range(x, *fit_range)]
+    yu        = poisson_sigma(y)
+    fseed     = (amp, mu, std)
+
+    par, err = par_and_err_from_seed(seed)
+    fr = FitResult(par = par,
+                   err = err,
+                   chi2 = NN,
+                   valid = False)
+    fp = None
+
+    with warnings.catch_warnings():
+        warnings.filterwarnings('error')  # in order to handle fit failures here
+        try:
+            fp, fr = gfit(x, y, yu, fseed)
+        except RuntimeWarning:   # this is the most usual failure, and usually solved trying fitx
+                                 # with a different seed
+            print(f' fit failed for seed  = {seed}, due to RunTimeWarning, retry fit ')
+            fseed = (10*fseed[0], fseed[1], fseed[2] )
+            try:
+                fp, fr = gfit(x, y, yu, fseed)
+            except RuntimeWarning: #  Give up on second failure
+                print(f' fit failed for seed  = {seed}, due to RunTimeWarning, give up ')
+        except OptimizeWarning:
+            print(f' OptimizeWarning was raised for seed  = {seed} due to OptimizeWarning')
+        except RuntimeError:
+            print(f' fit failed for seed  = {seed}  due to RunTimeError')
+        except TypeError:
+            print(f' fit failed for seed  = {seed}  due to TypeError')
+
+    return fp, fr
+
+
+def fit_energy(e : np.array,
+               nbins   : int,
+               range   : Tuple[float],
+               n_sigma : float = 3.0)->FitCollection:
+    """
+    Takes an "energy vector" (e.g, 1d array), with number of bins enbins and range erange, then:
+        1. Computes the histogram of e with enbins in erange. This returns an array of bin
+        edges (b), and bin contents (y). The array (b) is shifted to bin centers (x)
+        2. The arrays x and y are fitted to a gaussian, in a range given by an interval
+        arround the estimation of the maximum of the gaussian. The interval size is estimated
+        by multiplying n_sigma by the estimation of the gaussian std.
+
+    The result of the fit is a fit collection, that includes a FitPar and a HistoPar objects
+    needed for printing and plotting the fit result.
+       """
+
+    y, b = np.histogram(e, bins= nbins, range=range)
+    x = shift_to_bin_centers(b)
+    bin_size = (range[1] - range[0]) / nbins
+    seed = gaussian_parameters(e, range, bin_size)
+
+    fp, fr = gaussian_fit(x, y, seed, n_sigma)
+
+    hp = HistoPar(var      = e,
+                  nbins    = nbins,
+                  range    = range)
+
+    return FitCollection(fp = fp, hp = hp, fr = fr)
+
+
+
+def fit_gaussian_experiments(exps    : np.array,
+                             nbins   : int       = 50,
+                             range   : Range     = (9e+3, 11e+3),
+                             n_sigma : int       =3)->List[FitCollection]:
+    return [fit_energy(e, nbins, range, n_sigma) for e in exps]
+
+
+
+def par_and_err_from_seed(seed : GaussPar) ->Tuple[np.array, np.array]:
+    par = np.zeros(3)
+    err = np.zeros(3)
+    par[0] = seed.amp.value
+    par[1] = seed.mu.value
+    par[2] = seed.std.value
+    err[0] = seed.amp.uncertainty
+    err[1] = seed.mu.uncertainty
+    err[2] = seed.std.uncertainty
+    return par, err
+def gaussian_params_from_fcs(fcs : FitCollection) ->Iterable[float]:
+    mus   = np.array([fc.fr.par[1] for fc in fcs])
+    umus  = np.array([fc.fr.err[1] for fc in fcs])
+    stds  = np.array([fc.fr.par[2] for fc in fcs])
+    ustds = np.array([fc.fr.err[2] for fc in fcs])
+    chi2s = np.array([fc.fr.chi2   for fc in fcs])
+    return mus, umus, stds, ustds, chi2s
