+++ krcal/NB_utils/fit_energy_functions_test.py
@@ -4,33 +4,20 @@ Tests for fit_functions
 
 import numpy as np
 
-from pytest        import mark
+
 from pytest        import approx
-from pytest        import raises
-from flaky         import flaky
-from numpy.testing import assert_array_equal
-from numpy.testing import assert_allclose
 
 from hypothesis            import given, settings
-from hypothesis.strategies import integers
 from hypothesis.strategies import floats
-from invisible_cities.core.testing_utils       import exactly
-from invisible_cities.core.testing_utils       import float_arrays
-from invisible_cities.core.testing_utils       import FLOAT_ARRAY
-from invisible_cities.core.testing_utils       import random_length_float_arrays
-from invisible_cities.icaro. hst_functions     import shift_to_bin_centers
-
-from invisible_cities.core .stat_functions     import poisson_sigma
-from invisible_cities.core.core_functions      import in_range
 from invisible_cities.evm  .ic_containers      import Measurement
 
 from . fit_energy_functions       import gaussian_parameters
 from . fit_energy_functions       import fit_energy
-from . fit_functions              import chi2
-from . stat_functions             import mean_and_std
-from . stat_functions             import gaussian_experiments
+from ..core. stat_functions             import mean_and_std
+from ..core. testing_utils             import gaussian_experiments
 from . fit_energy_functions       import fit_gaussian_experiments
 from . fit_energy_functions       import gaussian_params_from_fcs
+
 
 
 @given(floats(min_value = -100,
